+++ src/main/kotlin/ru/marathontracker/gpd/models/responses/authorization/AuthorizationResponses.kt
@@ -22,8 +22,8 @@ data class AuthenticateResponse(
     val name: String,
     val age: Int,
     val sex: Sex,
-    val height: Float,
-    val weight: Float,
+    val height: Int,
+    val weight: Int,
     val phone: String,
 )
